+++ packages/editor/renderer/src/__tests__/unit/email/index.ts
@@ -19,6 +19,8 @@ import * as tableNumberedColumn from '../../__fixtures__/table-numbered-column.a
 import * as layoutColumnSection from '../../__fixtures__/layout-column-section.adf.json';
 import * as extensions from '../../__fixtures__/extensions.adf.json';
 import * as date from '../../__fixtures__/date.adf.json';
+import * as mediaSingle from '../../__fixtures__/media-single.adf.json';
+import * as mediaGroup from '../../__fixtures__/media-group.adf.json';
 
 import * as image from '../../__fixtures__/image.adf.json';
 import * as placeholder from '../../__fixtures__/placeholder.adf.json';
@@ -58,6 +60,16 @@ describe('Renderer - EmailSerializer', () => {
 
   it('should apply no mark for breakout marks', () => {
     const output = render(breakout);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render media single correctly', () => {
+    const output = render(mediaSingle);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render media group correctly', () => {
+    const output = render(mediaGroup);
     expect(output).toMatchSnapshot();
   });