+++ packages/elements/emoji/src/util/analytics.ts
@@ -6,7 +6,7 @@ import {
 import {
   name as packageName,
   version as packageVersion,
-} from '../package.json';
+} from '../version.json';
 
 export const createAndFireEventInElementsChannel: CreateAndFireEventFunction = createAndFireEvent(
   'fabric-elements',