+++ packages/navigation/atlassian-switcher/examples/10-jira-app-switcher-example.tsx
@@ -15,7 +15,27 @@ class JiraSwitcherExample extends React.Component {
   }
 
   openDrawer = () => {
-    mockEndpoints('jira', undefined, REQUEST_FAST);
+    mockEndpoints(
+      'jira',
+      originalMockData => {
+        return {
+          ...originalMockData,
+          LICENSE_INFORMATION_DATA: {
+            hostname: 'https://some-random-instance.atlassian.net',
+            firstActivationDate: 1492488658539,
+            maintenanceEndDate: '2017-04-24',
+            maintenanceStartDate: '2017-04-17',
+            products: {
+              'jira-software.ondemand': {
+                billingPeriod: 'ANNUAL',
+                state: 'ACTIVE',
+              },
+            },
+          },
+        };
+      },
+      REQUEST_FAST,
+    );
     this.setState({
       isDrawerOpen: true,
     });