+++ packages/editor/conversation/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/conversation
 
+## 14.1.0
+- [minor] [33563e651d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/33563e651d):
+
+  - Add support for user-state lozenges
+
+## 14.0.2
+- Updated dependencies [c2c36de22b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c2c36de22b):
+  - @atlaskit/editor-common@36.0.0
+  - @atlaskit/editor-core@107.0.0
+  - @atlaskit/renderer@45.0.0
+  - @atlaskit/comment@7.0.4
+  - @atlaskit/reactions@16.1.3
+  - @atlaskit/util-data-test@11.1.5
+
 ## 14.0.1
 - [patch] [1bcaa1b991](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1bcaa1b991):