+++ packages/media/media-card/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/media-card
 
+## 59.1.0
+- [minor] [e1c1fa454a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e1c1fa454a):
+
+  - Support external image identifier in MediaViewer
+
+## 59.0.0
+- Updated dependencies [c2c36de22b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c2c36de22b):
+  - @atlaskit/media-viewer@40.0.0
+  - @atlaskit/media-store@11.0.3
+  - @atlaskit/media-test-helpers@21.1.0
+  - @atlaskit/media-core@29.1.0
+
 ## 58.0.1
 - [patch] [106d046114](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/106d046114):